Add esc keypress back feature , Breakdown all the component to different files and called them
