+++ src/components/Box.js
@@ -1,0 +1,17 @@
+import PropTypes from "prop-types";
+import { useState } from "react";
+
+export default function Box({ children }) {
+  const [isOpen, setIsOpen] = useState(true);
+  return (
+    <div className="box">
+      <button className="btn-toggle" onClick={() => setIsOpen((open) => !open)}>
+        {isOpen ? "–" : "+"}
+      </button>
+      {isOpen && children}
+    </div>
+  );
+}
+Box.propTypes = {
+  children: PropTypes.func,
+};
