+++ src/components/ErrorMessage.js
@@ -1,0 +1,7 @@
+export default function ErrorMessage({ message }) {
+  return (
+    <p className="error">
+      <span>⛔</span> {message}
+    </p>
+  );
+}
