+++ src/components/Loader.js
@@ -1,0 +1,3 @@
+export default function Loader() {
+  return <p className="loader">Loading...</p>;
+}
