+++ src/components/Logo.js
@@ -1,0 +1,10 @@
+export default function Logo() {
+  return (
+    <div className="logo">
+      <span role="img" aria-label="Popcorn">
+        🍿
+      </span>
+      <h1>usePopcorn</h1>
+    </div>
+  );
+}
