+++ src/components/Main.js
@@ -1,0 +1,8 @@
+import PropTypes from "prop-types";
+
+export default function Main({ children }) {
+  return <main className="main">{children}</main>;
+}
+Main.propTypes = {
+  children: PropTypes.func,
+};
