+++ src/globalConfig/movieApi-key.js
@@ -1,0 +1,5 @@
+const APIKEY = {
+  KEY: "e799a191",
+};
+
+module.exports = APIKEY;
